accel working once from userland

+++ userland/examples/tests/FXOS8700CQ/main.c
@@ -26,8 +26,8 @@ int main() {
 		// uint16_t accel_y = 1; 
 		// uint16_t accel_z = 1; 
 
-		double accel_mag = FXOS8700CQ_read_accel_mag(); //sqrt(accel_x * accel_x + accel_y * accel_y + accel_z * accel_z); 
-		printf("accel mag = %f\n", accel_mag); 
+		unsigned accel_mag = FXOS8700CQ_read_accel_mag(); //sqrt(accel_x * accel_x + accel_y * accel_y + accel_z * accel_z); 
+		printf("accel mag = %u\n", accel_mag); 
 		accel_mags[ii] = accel_mag + g;
 		delay_ms(500); 
 	}
